requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2
pydantic
typing
logging
azure-storage-queue==12.11.0
azure-storage-blob==12.22.0
aiohttp
